+++ src/modules/contacts/types/Contact.types.ts
@@ -5,17 +5,9 @@ import type {
 } from '@webitel/api-services/gen/models';
 
 import type { ServiceConfigurable } from '../../configs';
-import type {
-	MessageSendTextParams,
-	MessageSendTextRawResponse,
-} from '../../messages/types/Message.types';
+import type { IMessageSender } from '../../messages/types/Message.types';
 
-interface IContact extends ContactModel, ServiceConfigurable {
-	sendTextMessage: (
-		body: string,
-		params?: Omit<MessageSendTextParams, 'body' | 'to'>,
-	) => Promise<MessageSendTextRawResponse>;
-}
+interface IContact extends ContactModel, ServiceConfigurable, IMessageSender {}
 
 type ContactSearchResult = Omit<ContactSearchRawResult, 'items'> & {
 	items: IContact[];